+++ src/pygaps/calculations/models_hk.py
@@ -1,4 +1,7 @@
-"""Contains dictionaries for use in the Horvath-Kawazoe method."""
+"""
+Dictionaries or generators which provide properties
+for use in the Horvath-Kawazoe method.
+"""
 
 from ..utilities.exceptions import ParameterError
 
@@ -9,16 +12,22 @@ PROPERTIES_CARBON = {
     'surface_density': 3.845E19,           # molecules/m2
 }
 
-PROPERTIES_OXIDE_ION = {
+PROPERTIES_AlSi_OXIDE_ION = {
     'molecular_diameter': 0.276,            # nm
     'polarizability': 2.5E-3,               # nm3
     'magnetic_susceptibility': 1.3E-7,      # nm3
     'surface_density': 1.315E19,            # molecules/m2
 }
-
+PROPERTIES_AlPh_OXIDE_ION = {
+    'molecular_diameter': 0.260,            # nm
+    'polarizability': 2.5E-3,               # nm3
+    'magnetic_susceptibility': 1.3E-7,      # nm3
+    'surface_density': 1.000E19,            # molecules/m2
+}
 
 _ADSORBENT_MODELS = {'Carbon(HK)': PROPERTIES_CARBON,
-                     'OxideIon(SF)': PROPERTIES_OXIDE_ION}
+                     'AlSiOxideIon': PROPERTIES_AlSi_OXIDE_ION,
+                     'AlPhOxideIon': PROPERTIES_AlPh_OXIDE_ION}
 
 
 def get_hk_model(model):
@@ -47,11 +56,11 @@ def get_hk_model(model):
     if isinstance(model, str):
         if model not in _ADSORBENT_MODELS:
             raise ParameterError(
-                "Model {} not an option for pore size distribution.".format(
+                "Model parameters {} not an option for pore size distribution.".format(
                     model),
-                "Available models are {}".format(_ADSORBENT_MODELS.keys()))
-        else:
-            a_model = _ADSORBENT_MODELS[model]
+                "Available model parameters are {}".format(_ADSORBENT_MODELS.keys()))
+
+        return _ADSORBENT_MODELS[model]
 
     # If the model is an dictionary, use it as is
     elif isinstance(model, dict):
@@ -63,7 +72,7 @@ def get_hk_model(model):
                     'in the units of {}'.format(key[0], key[1])
                 )
 
-        a_model = model
+        return model
 
     # Raise error if anything else is passed
     else:
@@ -71,5 +80,3 @@ def get_hk_model(model):
             "Not an option for pore size distribution.",
             "Available models are {}".format(_ADSORBENT_MODELS.keys()),
             "or pass a dictionary with the required parameters")
-
-    return a_model