+++ src/pygaps/calculations/psd_microporous.py
@@ -1,5 +1,6 @@
 """
-Module contains 'classical' methods of calculating a pore size distribution for pores in
+Module contains 'classical' methods of calculating
+a pore size distribution for pores in
 the micropore range (<2 nm).
 """
 
@@ -7,7 +8,122 @@ import numpy
 import scipy.constants as const
 import scipy.optimize as opt
 
+from ..classes.adsorbate import Adsorbate
+from ..graphing.calcgraph import psd_plot
 from ..utilities.exceptions import ParameterError
+from .models_hk import get_hk_model
+
+_MICRO_PSD_MODELS = ['HK']
+_PORE_GEOMETRIES = ['slit', 'cylinder', 'sphere']
+
+
+def psd_microporous(isotherm,
+                    psd_model='HK',
+                    pore_geometry='slit',
+                    branch='ads',
+                    adsorbent_model='Carbon(HK)',
+                    adsorbate_model=None,
+                    verbose=False):
+    """
+    Calculate the microporous size distribution using a 'classical' model.
+
+    Parameters
+    ----------
+    isotherm : Isotherm
+        Isotherm for which the pore size distribution will be calculated.
+    psd_model : str
+        The pore size distribution model to use.
+    pore_geometry : str
+        The geometry of the adsorbent pores.
+    branch : {'ads', 'des'}, optional
+        Branch of the isotherm to use. It defaults to adsorption.
+    adsorbent_model : str or dict
+        The adsorbent model to use for PSD, It defaults to Carbon(HK).
+    adsorbate_model : dict or `None`
+        The adsorbate properties to use for PSD, If null, properties are
+        automatically searched from the Adsorbent.
+    verbose : bool
+        Prints out extra information on the calculation and graphs the results.
+
+    Returns
+    -------
+    dict
+        A dictionary with the pore widths and the pore distributions, of the form:
+
+            - ``pore_widths`` (array) : the widths of the pores
+            - ``pore_distribution`` (array) : contribution of each pore width to the
+              overall pore distribution
+
+    Notes
+    -----
+    Calculates the pore size distribution using a 'classical' model which attempts to
+    describe the adsorption in a pore of specific width w at a relative pressure p/p0
+    as a single function :math:`p/p0 = f(w)`. This function uses properties of the
+    adsorbent and adsorbate as a way of determining the pore size distribution.
+
+    Currently, the methods provided are:
+
+        - the HK, or Horvath-Kawazoe method
+
+    A common mantra of data processing is: "garbage in = garbage out". Only use methods
+    when you are aware of their limitations and shortcomings.
+
+    See Also
+    --------
+    pygaps.calculations.psd_microporous.psd_horvath_kawazoe : low level HK (Horvath-Kawazoe) method
+
+    """
+    # Function parameter checks
+    if psd_model is None:
+        raise ParameterError("Specify a model to generate the pore size"
+                             " distribution e.g. psd_model=\"HK\"")
+    if psd_model not in _MICRO_PSD_MODELS:
+        raise ParameterError("Model {} not an option for psd.".format(psd_model),
+                             "Available models are {}".format(_MICRO_PSD_MODELS))
+    if pore_geometry not in _PORE_GEOMETRIES:
+        raise ParameterError(
+            "Geometry {} not an option for pore size distribution.".format(
+                pore_geometry),
+            "Available geometries are {}".format(_PORE_GEOMETRIES))
+    if not isinstance(isotherm.adsorbate, Adsorbate):
+        raise ParameterError("Isotherm adsorbate is not known, cannot calculate PSD.")
+
+    # Get adsorbate properties
+    if adsorbate_model is None:
+        adsorbate_model = {
+            'molecular_diameter': isotherm.adsorbate.get_prop('molecular_diameter'),
+            'polarizability': isotherm.adsorbate.get_prop('polarizability'),
+            'magnetic_susceptibility': isotherm.adsorbate.get_prop('magnetic_susceptibility'),
+            'surface_density': isotherm.adsorbate.get_prop('surface_density'),
+            'liquid_density': isotherm.adsorbate.liquid_density(isotherm.t_iso),
+            'adsorbate_molar_mass': isotherm.adsorbate.molar_mass(),
+        }
+
+    # Get adsorbent properties
+    adsorbent_properties = get_hk_model(adsorbent_model)
+
+    # Read data in
+    loading = isotherm.loading(branch=branch,
+                               loading_basis='molar',
+                               loading_unit='mmol')
+    pressure = isotherm.pressure(branch=branch,
+                                 pressure_mode='relative')
+
+    # Call specified pore size distribution function
+    if psd_model == 'HK':
+        pore_widths, pore_dist, pore_vol_cum = psd_horvath_kawazoe(
+            loading, pressure, isotherm.t_iso, pore_geometry,
+            adsorbate_model, adsorbent_properties)
+
+    if verbose:
+        psd_plot(pore_widths, pore_dist,
+                 pore_vol_cum=pore_vol_cum, log=False, right=2.5, method=psd_model)
+
+    return {
+        'pore_widths': pore_widths,
+        'pore_distribution': pore_dist,
+        'pore_volume_cumulative': pore_vol_cum,
+    }
 
 
 def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
@@ -49,6 +165,8 @@ def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
         The widths of the pores.
     pore_dist : array
         The distributions for each width.
+    pore_vol_cum : array
+        Cumulative pore volume.
 
     Notes
     -----
@@ -139,9 +257,9 @@ def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
     if pore_geometry == 'slit':
         pass
     elif pore_geometry == 'cylinder':
-        raise NotImplementedError
+        raise NotImplementedError('Currently only slit pores supported.')
     elif pore_geometry == 'sphere':
-        raise NotImplementedError
+        raise NotImplementedError('Currently only slit pores supported.')
 
     if adsorbent_properties is None:
         raise ParameterError(
@@ -178,36 +296,32 @@ def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
             "{}".format(' '.join(missing)))
 
     # dictionary unpacking
-    d_adsorbate = adsorbate_properties.get('molecular_diameter')
-    d_adsorbent = adsorbent_properties.get('molecular_diameter')
-    p_adsorbate = adsorbate_properties.get('polarizability') * 1e-27            # to m3
-    p_adsorbent = adsorbent_properties.get('polarizability') * 1e-27            # to m3
-    m_adsorbate = adsorbate_properties.get('magnetic_susceptibility') * 1e-27   # to m3
-    m_adsorbent = adsorbent_properties.get('magnetic_susceptibility') * 1e-27   # to m3
-    n_adsorbate = adsorbate_properties.get('surface_density')
-    n_adsorbent = adsorbent_properties.get('surface_density')
+    d_gas = adsorbate_properties.get('molecular_diameter')
+    d_mat = adsorbent_properties.get('molecular_diameter')
+    p_gas = adsorbate_properties.get('polarizability') * 1e-27            # to m3
+    p_mat = adsorbent_properties.get('polarizability') * 1e-27            # to m3
+    m_gas = adsorbate_properties.get('magnetic_susceptibility') * 1e-27   # to m3
+    m_mat = adsorbent_properties.get('magnetic_susceptibility') * 1e-27   # to m3
+    n_gas = adsorbate_properties.get('surface_density')
+    n_mat = adsorbent_properties.get('surface_density')
     liquid_density = adsorbate_properties.get('liquid_density')
     adsorbate_molar_mass = adsorbate_properties.get('adsorbate_molar_mass')
 
     # calculation of constants and terms
     e_m = const.electron_mass
     c_l = const.speed_of_light
-    effective_diameter = d_adsorbate + d_adsorbent
+    effective_diameter = d_gas + d_mat
     sigma = (2 / 5)**(1 / 6) * effective_diameter / 2
     sigma_si = sigma * 1e-9
 
-    a_adsorbent = 6 * e_m * c_l ** 2 * p_adsorbate * p_adsorbent /\
-        (p_adsorbate / m_adsorbate + p_adsorbent / m_adsorbent)
-    a_adsorbate = 3 * e_m * c_l**2 * p_adsorbate * m_adsorbate / 2
+    a_mat = 6 * e_m * c_l ** 2 * p_gas * p_mat / (p_gas / m_gas + p_mat / m_mat)
+    a_gas = 3 * e_m * c_l**2 * p_gas * m_gas / 2
 
-    constant_coefficient = const.Avogadro / \
-        (const.gas_constant * temperature) * \
-        (n_adsorbate * a_adsorbate + n_adsorbent * a_adsorbent) / \
-        (sigma_si**4)
+    constant_coefficient = const.Avogadro / (const.gas_constant * temperature) * \
+        (n_gas * a_gas + n_mat * a_mat) / (sigma_si**4)
 
-    constant_interaction_term = - \
-        ((sigma**4) / (3 * (effective_diameter / 2)**3) -
-         (sigma**10) / (9 * (effective_diameter / 2)**9))
+    constant_interaction_term = - ((sigma**4) / (3 * (effective_diameter / 2)**3) -
+                                   (sigma**10) / (9 * (effective_diameter / 2)**9))
 
     def h_k_pressure(l_pore):
         pressure = numpy.exp(constant_coefficient / (l_pore - effective_diameter) *
@@ -224,7 +338,7 @@ def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
         def h_k_minimization(l_pore):
             return numpy.abs(h_k_pressure(l_pore) - p_point)
         res = opt.minimize_scalar(h_k_minimization)
-        p_w.append(res.x - d_adsorbent)
+        p_w.append(res.x - d_mat)
 
     # finally calculate pore distribution
     pore_widths = numpy.array(p_w)
@@ -232,4 +346,4 @@ def psd_horvath_kawazoe(loading, pressure, temperature, pore_geometry,
     volume_adsorbed = loading * adsorbate_molar_mass / liquid_density / 1000    # cm3/g
     pore_dist = numpy.diff(volume_adsorbed) / numpy.diff(pore_widths)
 
-    return avg_pore_widths, pore_dist
+    return avg_pore_widths, pore_dist, volume_adsorbed[1:]