+++ tests/calculations/test_models_hk.py
@@ -1,5 +1,11 @@
 """
-This test module has tests relating to Horvath-Kawazoe models
+Tests relating to Horvath-Kawazoe model validation.
+
+All functions in /calculations/models_hk.py are tested here.
+The purposes are:
+
+    - testing that the "function getter" is performing as expected.
+
 """
 
 import pytest
@@ -10,21 +16,20 @@ from pygaps.utilities.exceptions import ParameterError
 
 @pytest.mark.characterisation
 class TestHKModels():
-    """
-    Tests the HK models
-    """
+    """Tests the HK models."""
 
-    def test_hk_model_string(self):
-        "Just gets stored models"
+    def test_get_hk_model_string(self):
+        """Just gets stored models."""
         for model in hk._ADSORBENT_MODELS:
             assert hk.get_hk_model(model) == hk._ADSORBENT_MODELS[model]
 
+    def test_get_hk_model_error(self):
+        """Check if errors are raised."""
         with pytest.raises(ParameterError):
             hk.get_hk_model('bad_model')
 
-    def test_hk_model_dict(self):
-        "Passes a dict model"
-
+    def test_get_hk_model_dict(self):
+        """When passing a dict, we check for consistency and return the same dict."""
         model_dict = dict(
             molecular_diameter=0.276,            # nm
             polarizability=2.5E-3,              # nm3
@@ -34,12 +39,8 @@ class TestHKModels():
 
         assert hk.get_hk_model(model_dict) == model_dict
 
+        # dictionary parameters should be checked
         model_dict.pop('molecular_diameter')
 
         with pytest.raises(ParameterError):
             hk.get_hk_model(model_dict)
-
-    def test_hk_model_error(self):
-        "Error when nothing is passed"
-        with pytest.raises(ParameterError):
-            hk.get_hk_model(None)