+++ tests/conftest.py
@@ -1,7 +1,6 @@
 """
 Configuration file for pytest and commonly used fixtures
 """
-import matplotlib.pyplot as plt
 import pandas
 import pytest
 
@@ -22,20 +21,6 @@ def pytest_runtest_setup(item):
         previousfailed = getattr(item.parent, "_previousfailed", None)
         if previousfailed is not None:
             pytest.xfail("previous test failed (%s)" % previousfailed.name)
-
-# matplotlib functionality
-
-
-@pytest.fixture(scope='function')
-def noplot():
-    plt.ioff()
-    return
-
-
-@pytest.fixture(scope='function')
-def doplot():
-    plt.ion()
-    return
 
 # Global fixtures